requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
httpx>=0.24.1
pandas>=2.1.0
pyarrow>=14.0.1
pydantic>=2.4.0
python-dotenv>=1.0.0
streamlit>=1.29.0
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0
requests>=2.31.0
tensorflow>=2.15.0
